requirements.txt
Flask
flask-socketio
flask-cors
python-dotenv
requests
boto3
langchain
langchain-google-genai
google-generativeai
faiss-cpu
numpy
eventlet
gunicorn
gevent
gevent-websocket